backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
numpy==2.1.0
scikit-learn==1.5.2
python-multipart==0.0.9
